refactor(rules): split ruleaction builtin registration by domain
